src/mqtt_bridge/dev-requirements.txt
# thanks to rospypi project, we can install ros packages in pure python
# See: https://github.com/rospypi/simple
--extra-index-url https://rospypi.github.io/simple/
catkin-pkg
geometry-msgs
inject>=4.0
mock
msgpack-python>=0.4.8
paho-mqtt>=1.2
pymongo
pytest
pyyaml
git+https://github.com/RobotWebTools/rosbridge_suite.git#subdirectory=rosbridge_library
#git clone --branch ros1 --single-branch https://github.com/RobotWebTools/rosbridge_suite.git --depth 1 --recurse-submodules rosbridge_suite_ros1
rosgraph
rosgraph-msgs
roslib
rosmsg
rospy<1.15  # 1.15 is for python3
rostest
sensor-msgs
std-msgs
